requirements.txt
# GenieGuard: World-Sim CI - Dependencies

# Browser automation
playwright>=1.40.0

# Google Gemini API (for LLM patch selection)
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
